fix: change view. Refs #179
